Feat: Added the refactored field-type component
Added in the first component into the new turborepo. Eslint is also setup to work with the component but currently throws errors due to the current use of any in the component.

I have also cleaned up the repo a bit, getting rid of some of the old turbo repo files that we no longer need.